% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met2model.RothC.R
\name{met2model.RothC}
\alias{met2model.RothC}
\title{Extract monthly weather from CF file for input to RothC}
\usage{
met2model.RothC(
  in.path,
  in.prefix,
  outfolder,
  start_date,
  end_date,
  overwrite = FALSE
)
}
\arguments{
\item{in.path}{path on disk where CF files live}

\item{in.prefix}{prefix for each file}

\item{outfolder}{location where model specific output is written.}

\item{start_date, end_date}{When to start and end output.
Specify as exact dates, but output will be padded to whole months.}

\item{overwrite}{logical: replace output files if they already exist?}
}
\value{
data frame summarizing file metadata
}
\description{
Input files need to be named `<in.path>/<in.prefix>.YYYY.nc`
}
\details{
Output files are named `<outfolder>/<in.prefix>.YY-mm.YY-mm.dat`
with one line per month and columns for temperature, rainfall,
and evaporation.

Note that the created file contains only weather data and not any of the
soil or management data needed for RothC's single combined input file.
See `write.config.RothC()` for assembly into a model-ready RothC_input.dat`.
}
\author{
Chris Black
}
