% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model2netcdf.RothC.R
\name{model2netcdf.RothC}
\alias{model2netcdf.RothC}
\title{Convert RothC output to PEcAn-formatted netCDF}
\usage{
model2netcdf.RothC(outdir, sitelat, sitelon, start_date, end_date)
}
\arguments{
\item{outdir}{Location of model output}

\item{sitelat}{Latitude of the site}

\item{sitelon}{Longitude of the site}

\item{start_date}{Start time of the simulation}

\item{end_date}{End time of the simulation}
}
\description{
Convert RothC output to PEcAn-formatted netCDF
}
\author{
Chris Black
}
