% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PEcAn.RothC-package.R
\docType{package}
\name{PEcAn.RothC-package}
\alias{PEcAn.RothC}
\alias{PEcAn.RothC-package}
\title{PEcAn.RothC-package}
\description{
PEcAn support for the RothC soil carbon model
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pecanproject.github.io}
  \item Report bugs at \url{https://github.com/PecanProject/pecan/issues}
}

}
\author{
\strong{Maintainer}: Chris Black \email{chris@ckblack.org}

Other contributors:
\itemize{
  \item Rothamsted Research [copyright holder]
}

}
\keyword{internal}
