% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_restart.RothC.R
\name{read_restart.RothC}
\alias{read_restart.RothC}
\title{Read restart template for SDA}
\usage{
read_restart.RothC(outdir, runid, stop.time, settings, var.names, params)
}
\arguments{
\item{outdir}{Output directory}

\item{runid}{Run ID}

\item{stop.time}{Year that is being read}

\item{settings}{PEcAn settings object}

\item{var.names}{Variable names to be extracted}

\item{params}{Any parameters required for state calculations}
}
\value{
Forecast numeric matrix
}
\description{
Reads restart files from RothC. Not implemented yet.
}
\author{
Chris Black
}
