% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_soil_physics.R
\name{read_soil_physics}
\alias{read_soil_physics}
\title{Read soil parameters from a PEcAn soil physics file}
\usage{
read_soil_physics(path, model_depth = 23)
}
\arguments{
\item{path}{filepath to a netcdf,
probably read from `settings$run$inputs$soil_physics$path`}

\item{model_depth}{Soil depth to be simulated, in cm}
}
\value{
one-row dataframe with columns
  `depth_cm`, `clay_pct`, `silt_pct`,
  `bulkdens_g_cm3`, `org_C_pct`, `iom_tC_ha`
}
\description{
Reads values from a netCDF that follows the PEcAn soil standard,
  aggregates layers across the depth to be simulated,
  and converts to names and units expected by RothC.
}
\details{
TODO: Currently drops all of any layer that extends below `model_depth`.
  It would be better instead to weight all layers by their contribution
  to the requested depth.
}
