% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.config.RothC.R
\name{write.config.RothC}
\alias{write.config.RothC}
\title{Writes a RothC config file.}
\usage{
write.config.RothC(defaults, trait.values, settings, run.id)
}
\arguments{
\item{defaults}{list of defaults to process}

\item{trait.values}{vector of samples for a given trait}

\item{settings}{list of settings from pecan settings file}

\item{run.id}{id of run}
}
\value{
configuration file for MODEL for given run
}
\description{
Requires a pft xml object, a list of trait values for a single model run,
and the name of the file to create
}
\author{
Chris Black
}
