% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_restart.RothC.R
\name{write_restart.RothC}
\alias{write_restart.RothC}
\title{Write restart template for SDA}
\usage{
write_restart.RothC(
  outdir,
  runid,
  start.time,
  stop.time,
  settings,
  new.state,
  RENAME,
  new.params,
  inputs
)
}
\arguments{
\item{outdir}{outout directory}

\item{runid}{run id}

\item{start.time}{Time of current assimilation step}

\item{stop.time}{Time of next assimilation step}

\item{settings}{pecan settings list}

\item{new.state}{Analysis state matrix returned by \code{sda.enkf}}

\item{RENAME}{flag to either rename output file or not}

\item{new.params}{optional, additionals params to pass write.configs that
are deterministically related to the parameters updated by the analysis}

\item{inputs}{new input paths updated by the SDA workflow,
will be passed to write.configs}
}
\description{
Not implemented yet.
}
\author{
Chris Black
}
